Make `fetch-debuginfo` a standalone script
Move the `fetch-debuginfo` logic in a single python file which can be
invoked as a stand-alone script.
