+++ python/revng/internal/cli/_commands/fetch_debuginfo/__init__.py
@@ -2,42 +2,24 @@
 # This file is distributed under the MIT License. See LICENSE.md for details.
 #
 
-import os
-import struct
 from argparse import RawDescriptionHelpFormatter
 from pathlib import Path
-from typing import List
 
-from elftools.common.exceptions import ELFError
-from elftools.elf.elffile import ELFFile
-
-from revng.internal.cli.commands_registry import Command, CommandsRegistry, Options
-from revng.internal.support import configuration
+from revng.internal.cli.commands_registry import Command, CommandsRegistry
+from revng.internal.cli.commands_registry import Options as CLIOptions
+from revng.internal.support import cache_directory, configuration
 from revng.support import log_error
 
-from .common import log, logger
-from .elf import fetch_dwarf
-from .pe import fetch_pdb
-
-DEFAULT_ELF_SERVERS = ["https://debuginfo-cache.rev.ng/elf"]
-DEFAULT_PE_SERVERS = ["https://debuginfo-cache.rev.ng/pe"]
+from .common import Options, fetch_debuginfo, log, logger
 
 
-def log_success(file_path):
-    log(f"Result: {file_path}")
-
-
-def log_fail():
-    log("Result: No debug info found")
-
-
-def config_or_default(config_key: str, default: List[str]) -> List[str]:
+def get_config(config_key: str) -> list[str] | None:
     the_configuration = configuration()
-    if "debug-info-server-urls" not in the_configuration:
-        return default
-
-    if config_key not in the_configuration["debug-info-server-urls"]:
-        return default
+    if (
+        "debug-info-server-urls" not in the_configuration
+        or config_key not in the_configuration["debug-info-server-urls"]
+    ):
+        return None
 
     result = the_configuration["debug-info-server-urls"][config_key]
     assert isinstance(result, list)
@@ -45,68 +27,40 @@ def config_or_default(config_key: str, default: List[str]) -> List[str]:
     return result
 
 
-def is_pe(file):
-    file.seek(0, os.SEEK_SET)
-    mz_header = file.read(2)
-    if mz_header != b"MZ":
-        return False
-    file.seek(0x3C, os.SEEK_SET)  # 0x3C contains PE offset
-    pe_offset = struct.unpack("<i", file.read(4))[0]  # Offset is encoded as a little-endian int
-    file.seek(pe_offset, os.SEEK_SET)
-    return file.read(4) == b"PE\x00\x00"
-
-
-def is_elf(file):
-    file.seek(0, os.SEEK_SET)
-    return file.read(4) == b"\x7fELF"
-
-
+# This registers the `revng model fetch-debuginfo` command, while `common.py`
+# file can be used as a standalone script the revng counterpart has
+# revng-specific defaults (e.g. cache directory location) set.
 class FetchDebugInfoCommand(Command):
     def __init__(self):
         super().__init__(("model", "fetch-debuginfo"), "Fetch Debugging Information.")
-        self.elf_servers = config_or_default("dwarf", DEFAULT_ELF_SERVERS)
-        self.pe_servers = config_or_default("pe", DEFAULT_PE_SERVERS)
+        self.options = Options(output_dir=cache_directory())
+        if (config_elf_servers := get_config("dwarf")) is not None:
+            self.options.elf_servers = tuple(config_elf_servers)
+        if (config_pe_servers := get_config("pe")) is not None:
+            self.options.pe_servers = tuple(config_pe_servers)
 
     def register_arguments(self, parser):
         parser.formatter_class = RawDescriptionHelpFormatter
         parser.description = """Fetches debugging symbols from the internet.
-Remote urls can be overridden via REVNG_FETCH_DEBUGINFO_{ELF,PE}_SERVERS
+Remote urls can be overridden via revng.yml
 """
         parser.add_argument("input", help="The input file.")
 
-    def run(self, options: Options):
+    def run(self, options: CLIOptions):
         args = options.parsed_args
         logger.verbose = args.verbose
 
-        path = Path(args.input)
-        if not path.exists():
+        if not Path(args.input).exists():
             log_error("Could not find " + args.input)
             return 1
 
-        with path.open(mode="rb") as f:
-            # Check if it is an ELF.
-            if is_elf(f):
-                try:
-                    the_elffile = ELFFile(f)
-                    result = fetch_dwarf(the_elffile, path, self.elf_servers)
-                except ELFError as elf_error:
-                    log_error(str(elf_error))
-                    return 1
-            elif is_pe(f):
-                # TODO: handle _NT_SYMBOL_PATH and _NT_ALT_SYMBOL_PATH
-                # https://learn.microsoft.com/en-us/windows-hardware/drivers/debugger/symbol-path
-                result = fetch_pdb(path, self.pe_servers)
-            else:
-                log_fail()
-                return 1
-
+        result = fetch_debuginfo(args.input, self.options)
         if result is None:
-            # We have not found the debug info file.
-            log_fail()
+            log("Result: No debug info found")
             return 1
-
-        log_success(result)
-        return 0
+        else:
+            log(f"Result: {result}")
+            return 0
 
 
 def setup(commands_registry: CommandsRegistry):
